# openapi/swagger.yaml
openapi: 3.0.3
info:
  title: Twitter API clone with Node.js and TypeScript - OpenAPI 3.0
  description: |-
    This is a Twitter API clone with Node.js and TypeScript developed by Dao Anh Tuan. This API is built with Express.js, MongoDB, and TypeScript. The API provides functionalities such as user authentication, tweet creation, tweet interaction, media upload, bookmark, search, and chat conversation. For more information, please visit my [github repository](https://github.com/anhtuandao1311/twitter-api-nodejs-typescript) or contact me via email: <anhtuandao131102@gmail.com>. Thank you for your interest in my project 😊.
  version: 1.0.0
servers:
  - url: http://localhost:4000
...
paths:
...
  /users/me:
    get:
      tags:
        - users
      summary: Get current user profile
      description: Get current user profile
      operationId: getMe
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Get current user profile successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Get profile successfully
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
    patch:
      tags:
        - users
      summary: Update user profile
      description: Update user profile
      operationId: updateMe
      security:
        - BearerAuth: []
      requestBody:
        description: Update profile
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMeBody'
        required: true

      responses:
        '200':
          description: Update user profile successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Update profile successfully
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
  /users/refresh-token:
    post:
      tags:
        - users
      summary: Refresh access token
      description: Refresh access token
      operationId: refreshToken
      requestBody:
        description: Refresh token information
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
                  example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiNjRiZTBhZDJlNDNkMjQ2NDM5NGZlZWRiIiwidG9rZW5fdHlwZSI6MSwidmVyaWZ5IjoxLCJpYXQiOjE2OTEzODMyMjYsImV4cCI6MTcwMDAyMzIyNn0.bFnaBfxWl-Q48HCwAeMpNzZwTO9LEjcBd7Zyipjqr64
        required: true
      responses:
        '200':
          description: Refresh token successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Refresh token successfully
                  data:
                    $ref: '#/components/schemas/SuccessfulAuthentication'
  /users/resend-verify-email:
    post:
      tags:
        - users
      summary: Resend verify email
      description: Resend verify email
      operationId: resendVerifyEmail
      requestBody:
        description: Resend verify email information
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Resend verify email successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Resend verify email successfully
  /users/forgot-password:
    post:
      tags:
        - users
      summary: Forgot password
      description: Forgot password
      operationId: forgotPassword
      requestBody:
        description: Forgot password information
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: daoanhtuan123@gmail.com
        required: true
      responses:
        '200':
          description: Send forgot password email successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Send forgot password email successfully
...
  /users/reset-password:
    post:
      tags:
        - users
      summary: Reset password
      description: Reset password
      operationId: resetPassword
      requestBody:
        description: Reset password information
        content:
          application/json:
            schema:
              type: object
              properties:
                forgot_password_token:
                  type: string
                  example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiNjRiZTBhZDJlNDNkMjQ2NDM5NGZlZWRiIiwidG9rZW5fdHlwZSI6MSwidmVyaWZ5IjoxLCJpYXQiOjE2OTEzODMyMjYsImV4cCI6MTcwMDAyMzIyNn0.bFnaBfxWl-Q48HCwAeMpNzZwTO9LEjcBd7Zyipjqr64
                new_password:
                  type: string
                  example: DaoAnhTuan123!
                confirm_new_password:
                  type: string
                  example: DaoAnhTuan123!
        required: true
      responses:
        '200':
          description: Reset password successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Reset password successfully
...
  /users/change-password:
    put:
      tags:
        - users
      summary: Change password
      description: Change password
      operationId: changePassword
      security:
        - BearerAuth: []
      requestBody:
        description: Change password information
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
                  example: DaoAnhTuan123!
                new_password:
                  type: string
                  example: DaoAnhTuan12345!
                confirm_new_password:
                  type: string
                  example: DaoAnhTuan12345!
        required: true
      responses:
        '200':
          description: Change password successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Change password successfully
  /tweets:
    post:
      tags:
        - tweets
      summary: Create tweet
      description: Create tweet
      operationId: createTweet
      security:
        - BearerAuth: []
      requestBody:
        description: Tweet information
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TweetRequestBody'
        required: true
      responses:
        '201':
          description: Create tweet successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Create tweet successfully
...
  /bookmarks/create-bookmark:
    post:
      tags:
        - bookmarks
      summary: Create bookmark
      description: Create bookmark
      operationId: createBookmark
      security:
        - BearerAuth: []
      requestBody:
        description: Create bookmark information
        content:
          application/json:
            schema:
              type: object
              properties:
                tweet_id:
                  type: string
                  format: MongoId
                  example: 64be0ad2e43d2464394faedb
        required: true
      responses:
        '201':
          description: Create bookmark successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Create bookmark successfully
...
components:
  schemas:
...
    SuccessfulAuthentication:
      type: object
      properties:
        access_token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiNjRiZTBhZDJlNDNkMjQ2NDM5NGZlZWRiIiwidG9rZW5fdHlwZSI6MCwidmVyaWZ5IjoxLCJpYXQiOjE2OTEzODMyMjYsImV4cCI6MTY5MTQ2OTYyNn0.HTLX20cB7_z0c9c8FDg3MIx6RJEELHHlmJNZa94ku-o
        refresh_token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiNjRiZTBhZDJlNDNkMjQ2NDM5NGZlZWRiIiwidG9rZW5fdHlwZSI6MSwidmVyaWZ5IjoxLCJpYXQiOjE2OTEzODMyMjYsImV4cCI6MTcwMDAyMzIyNn0.bFnaBfxWl-Q48HCwAeMpNzZwTO9LEjcBd7Zyipjqr64
    User:
      type: object
      properties:
        _id:
          type: string
          format: MongoId
          example: 64be0ad2e43d2464394feedb
        name:
          type: string
          example: John Doe
        email:
          type: string
          example: johndoe@example.com
        date_of_birth:
          type: string
          format: ISO8601
          example: 2002-06-08T10:17:31.096Z
        created_at:
          type: string
          format: ISO8601
          example: 2024-05-05T12:00:00Z
        updated_at:
          type: string
          format: ISO8601
          example: 2024-05-05T12:00:00Z
        verify:
          $ref: '#/components/schemas/UserVerifyStatus'
        twitter_circle:
          type: array
          items:
            type: string
            format: MongoId
          example: ['64be0ad2e43d2464394feedb', '64be0ad2e43d2464394feedc']
        bio:
          type: string
          example: This is my bio.
        location:
          type: string
          example: San Francisco, CA
        website:
          type: string
          example: www.example.com
        username:
          type: string
          example: johndoe
        avatar:
          type: string
          example: http://localhost:4000/images/avatars/johndoe.jpg
        cover_photo:
          type: string
          example: http://localhost:4000/images/avatars/johndoe.jpg
    UserVerifyStatus:
      type: number
      enum: [Unverified, Verified, Banned]
      example: 1
...
    UpdateMeBody:
      type: object
      properties:
        name:
          type: string
          example: John Doe
        date_of_birth:
          type: string
          format: ISO8601
          example: 2000-06-08T10:17:31.096Z
        bio:
          type: string
          example: 'This is my bio.'
        location:
          type: string
          example: 'San Francisco, CA'
        website:
          type: string
          example: 'www.example.com'
        username:
          type: string
          example: 'johndoe'
        avatar:
          type: string
          example: 'http://localhost:4000/images/avatars/johndoe.jpg'
        cover_photo:
          type: string
          example: 'http://localhost:4000/images/avatars/johndoe.jpg'
    TweetRequestBody:
      type: object
      properties:
        type:
          $ref: '#/components/schemas/TweetType'
        audience:
          $ref: '#/components/schemas/TweetAudience'
        content:
          type: string
          description: Tweet content
          example: 'This is my tweet.'
        parent_id:
          type: string
          nullable: true
          description: ID of parent tweet
          example: 64be0ad2e43d2464394feedb
        hashtags:
          type: array
          items:
            type: string
          description: List of hashtags
          example: ['hashtag1', 'hashtag2']
        mentions:
          type: array
          items:
            type: string
          description: List of user IDs mentioned
          example: ['64be0ad2e43d2464394faedb', '64be0ad2e43d2464394faedc']
        media:
          type: array
          items:
            $ref: '#/components/schemas/Media'
          description: List of attached media
    TweetType:
      type: number
      enum: [Tweet, Retweet, Comment, QuoteTweet]
      example: 1
    TweetAudience:
      type: number
      enum: [Everyone, TwitterCircle]
      example: 1
...
    MediaType:
      type: number
      enum: [Image, Video]
      example: 1
    Media:
      type: object
      properties:
        url:
          type: string
          example: 'http://localhost:4000/images/avatars/johndoe.jpg'
        type:
          $ref: '#/components/schemas/MediaType'
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT